Initial refactor of cupertino/material/widgets app
