+++ packages/flutter/lib/src/cupertino/tab_view.dart
@@ -4,6 +4,7 @@
 
 import 'package:flutter/widgets.dart';
 
+import 'app.dart' show CupertinoApp;
 import 'route.dart';
 
 /// A single tab view with its own [Navigator] state and history.
@@ -114,7 +115,7 @@ class _CupertinoTabViewState extends State<CupertinoTabView> {
   @override
   void initState() {
     super.initState();
-    _heroController = HeroController(); // Linear tweening.
+    _heroController = CupertinoApp.createCupterinoTransitionController();
     _updateObservers();
   }
